Linted contracts

+++ contracts/interfaces/CERC20.sol
@@ -1,14 +1,14 @@
 pragma solidity 0.5.16;
 
 interface CERC20 {
-  function mint(uint mintAmount) external returns (uint);
-  function redeemUnderlying(uint redeemAmount) external returns (uint);
-  function borrow(uint borrowAmount) external returns (uint);
-  function repayBorrow(uint repayAmount) external returns (uint);
-  function borrowBalanceCurrent(address account) external returns (uint);
-  function exchangeRateCurrent() external returns (uint);
+    function mint(uint256 mintAmount) external returns (uint256);
+    function redeemUnderlying(uint256 redeemAmount) external returns (uint256);
+    function borrow(uint256 borrowAmount) external returns (uint256);
+    function repayBorrow(uint256 repayAmount) external returns (uint256);
+    function borrowBalanceCurrent(address account) external returns (uint256);
+    function exchangeRateCurrent() external returns (uint256);
 
-  function balanceOf(address account) external view returns (uint);
-  function decimals() external view returns (uint);
-  function underlying() external view returns (address);
+    function balanceOf(address account) external view returns (uint256);
+    function decimals() external view returns (uint256);
+    function underlying() external view returns (address);
 }
