+++ contracts/interfaces/KyberNetwork.sol
@@ -1,10 +1,19 @@
 pragma solidity 0.5.16;
 
 interface KyberNetwork {
-  function getExpectedRate(address src, address dest, uint srcQty) external view
-      returns (uint expectedRate, uint slippageRate);
+    function getExpectedRate(address src, address dest, uint256 srcQty)
+        external
+        view
+        returns (uint256 expectedRate, uint256 slippageRate);
 
-  function tradeWithHint(
-    address src, uint srcAmount, address dest, address payable destAddress, uint maxDestAmount,
-    uint minConversionRate, address walletId, bytes calldata hint) external payable returns(uint);
+    function tradeWithHint(
+        address src,
+        uint256 srcAmount,
+        address dest,
+        address payable destAddress,
+        uint256 maxDestAmount,
+        uint256 minConversionRate,
+        address walletId,
+        bytes calldata hint
+    ) external payable returns (uint256);
 }
